% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet_forest.R
\name{cleanup_distributed_forest}
\alias{cleanup_distributed_forest}
\title{Clean Up Distributed Forest}
\usage{
cleanup_distributed_forest(distributed_forest)
}
\arguments{
\item{distributed_forest}{A distributed forest object}
}
\description{
Properly cleans up resources used by distributed forest.
}
